Re-structure nulecule package.
